Create newGameController.js
Adding multiple controllers to handle different parts of the game. Each controller will be used for a specific template depending on what phase of the game they are in. This controller will be used at the very beginning of the game to help initialize players and create a new game.